requirements.txt
dataclasses-json>=0.5.2
decorator>=4.4.2
gcs-oauth2-boto-plugin>=2.7
gcsfs>=0.7.2
google-api-core>=1.26.1
google-api-python-client>=2.0.2
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.3
google-auth>=1.28.0
google-cloud-bigquery>=2.13.0
google-cloud-core>=1.6.0
google-cloud-firestore>=2.0.2
google-cloud-secret-manager>=2.8.0
google-cloud-storage>=1.36.2
google-reauth>=0.1.1
googleapis-common-protos>=1.53.0
httplib2>=0.19.0
oauth2client>=4.1.3
oauthlib>=3.1.0
pytz>=2021.1
requests-oauthlib>=1.3.0
requests>=2.27.1
